+++ src/printer/SMTLIB2Printer.cpp
@@ -213,7 +213,7 @@ void printVarDeclsToStream(ASTNodeSet& symbols, ostream& os)
   		{
   			string close = "";
 
-  			for (int i =0; i < c.size()-1; i++)
+  			for (long int i =0; i < (long int)c.size()-1; i++)
   			{
   				os << "(" << functionToSMTLIBName(kind,false);
   				os << " ";
